[refactor]: Service interface 제거
